Added better HTTP status support and replaced cookie header key

+++ src/main/java/com/rodaxsoft/todo/rest/TaskController.java
@@ -10,28 +10,29 @@
 */
 package com.rodaxsoft.todo.rest;
 
-import static com.rodaxsoft.todo.security.SecurityConstants.COOKIE_STRING;
+import static com.rodaxsoft.todo.security.SecurityConstants.HEADER_STRING;
 
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.CookieValue;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.rodaxsoft.todo.domain.Task;
+import com.rodaxsoft.todo.exception.ResourceNotFoundException;
+import com.rodaxsoft.todo.exception.ValidationException;
 import com.rodaxsoft.todo.service.ApplicationUserService;
 import com.rodaxsoft.todo.service.TaskService;
 import com.rodaxsoft.todo.validation.TaskCreateAndUpdateValidator;
-import com.rodaxsoft.todo.validation.ValidationException;
 
 /**
  * TaskController class
@@ -52,7 +53,7 @@ public class TaskController {
 	}
 
 	@PostMapping
-	public Task createTask(@RequestBody Task task, BindingResult result, @CookieValue(COOKIE_STRING) String token) {
+	public Task createTask(@RequestBody Task task, BindingResult result, @RequestHeader(HEADER_STRING) String token) {
 		taskValidator.validate(task, result);
 		if(result.hasErrors()) {
 			throw new ValidationException(result);
@@ -66,12 +67,16 @@ public class TaskController {
 	}
 	
 	@GetMapping
-	public List<Task> getTasks(@CookieValue(COOKIE_STRING) String token) {
+	public List<Task> getTasks(@RequestHeader(HEADER_STRING) String token) {
 		return taskService.getTasks();	
 	}
 	
 	@PutMapping(path = "/{id}")
 	public Task updateTask(@PathVariable Long id, @RequestBody Task updatedTask, BindingResult result) {
+		if(!taskService.exists(id)) {
+			throw new ResourceNotFoundException("Task not found");
+		}
+		
 		taskValidator.validate(updatedTask, result);
 		if (result.hasFieldErrors()) {
 			throw new ValidationException(result);
@@ -81,10 +86,14 @@ public class TaskController {
 		return taskService.updateTask(updatedTask);
 	}
 	
-	 @DeleteMapping(path = "/{id}", produces = MediaType.TEXT_PLAIN_VALUE)
-		public String deleteTask(@PathVariable Long id) {
-			taskService.deleteTask(id);
-			return "";
+	@DeleteMapping(path = "/{id}", produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
+	public String deleteTask(@PathVariable Long id) {
+		if(!taskService.exists(id)) {
+			throw new ResourceNotFoundException("Task not found");
 		}
+
+		taskService.deleteTask(id);
+		return "";
+	}
 
 }
